+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-## Oct 30 2018
-    * Added freeze method to produce setup or validation dataset
+## Nov 6 2018 - v0.2.0
+    * Added Dump method to create schema from existing database
+
+## Oct 30 2018 - v0.1.1
+    * Added freeze method to create setup or validation dataset from existing database
+
 
 ## Jul 1 2016 (Alpha)